add -spec support to rebar_ct command line
This change adds support for executing ct
test runs based on test specificiations,
which was missing previously. The rebar_ct
module now looks for any number of files
with a name ending in `test.spec` and if
it finds one or more, passes these after
the `-spec` argument to ct_run instead of
explicitly configuring the config, user
config and coverage config variables.

When no specifications are found, then the
module behaves as it did before this change,
and both the ct1 and (new) ct2 integration
tests appear to show this is a backwards
compatible patch.