Fix issue 500
This change resolves issue 500, by preventing
destructive commands (such as clean, compile
and create/template) from running inside any
dependency directory that is outside of the
current project base_dir.

The rationale for this alteration is that any
dependency installed outside of the base_dir,
was likely installed by some other means and
is therefore outside of rebar's area of
responsibility.

This fix also subsumes issue 763, which limits
the scope of the problem to the code:lib_dir
only. Whilst this is clearly a problem, it is
the author's view that any dependency that has
not been installed by rebar within a project
directory, should not be affected by its use.